TL Layer Doc refactoring
