+++ tensorlayer/layers/__init__.py
@@ -9,7 +9,6 @@ layers that allow user to apply ``tf.nn.lrn`` on ``network.outputs``.
 More functions can be found in `TensorFlow API <https://www.tensorflow.org/versions/master/api_docs/index.html>`__.
 """
 
-from .binary import *
 from .convolution import *
 from .core import *
 from .dense import *
@@ -25,8 +24,10 @@ from .normalization import *
 from .object_detection import *
 from .padding import *
 from .pooling import *
+from .quantize import *
 from .reconstruction import *
 from .recurrent import *
+from .scale import *
 from .shape import *
 from .spatial_transformer import *
 from .special_activation import *
